Updates db to change MachineChoice.value to a Floatfield, updates create_tree to only add to the db if an entry doesn't already exist.

+++ ttt/models.py
@@ -6,5 +6,5 @@ class MachinePath(models.Model):
 
 class MachineChoice(models.Model):
     option = models.CharField(max_length=2)
-    value = models.IntegerField()
+    value = models.FloatField()
     path = models.ForeignKey(MachinePath, on_delete= models.CASCADE)
